+++ packages/cli/src/zosfiles/download/dsm/DataSetMatching.handler.ts
@@ -72,6 +72,7 @@ export default class DataSetMatchingHandler extends ZosFilesBaseHandler {
             maxConcurrentRequests: commandParameters.arguments.maxConcurrentRequests,
             preserveOriginalLetterCase: commandParameters.arguments.preserveOriginalLetterCase,
             failFast: commandParameters.arguments.failFast,
+            overwrite: commandParameters.arguments.overwrite,
             task: downloadStatus,
             responseTimeout: commandParameters.arguments.responseTimeout
         };